+++ ui/src/Components/SilenceModal/Matchers.ts
@@ -1,12 +1,23 @@
 import { FormatQuery, QueryOperators, StaticLabels } from "Common/Query";
 import type { MultiValueOptionT } from "Common/Select";
-import { MatcherT, MatcherToOperator } from "Stores/SilenceFormStore";
+import {
+  MatcherT,
+  MatcherToOperator,
+  EscapeRegex,
+} from "Stores/SilenceFormStore";
 
 const MatcherToFilter = (matcher: MatcherT): string => {
+  const values = matcher.values.map((v) =>
+    v.wasCreated
+      ? v
+      : matcher.isRegex
+      ? { ...v, value: EscapeRegex(v.value) }
+      : v
+  );
   const value =
-    matcher.values.length > 1
-      ? `(${matcher.values.map((v) => v.value).join("|")})`
-      : matcher.values[0].value;
+    values.length > 1
+      ? `(${values.map((v) => v.value).join("|")})`
+      : values[0].value;
   return FormatQuery(
     matcher.name,
     MatcherToOperator(matcher),